Update for latest agents.
Also fix identify match counting to use longs (it was overflowing
integer range when calculating percentages).

+++ src/com/leafdigital/browserstats/identify/Identify.java
@@ -35,7 +35,7 @@ public class Identify extends CommandLineTool implements UserAgentReader.Handler
 
 	private TestType test = TestType.NONE;
 	private String[] testParams = null;
-	private int minCount, matched, unmatched, matchedCount, unmatchedCount;
+	private long minCount, matched, unmatched, matchedCount, unmatchedCount;
 
 	private IdentifyResults results;
